requirements.txt
# Core API framework
fastapi==0.116.1
uvicorn[standard]==0.34.0
pydantic==2.11.0

# ML dependencies
sentence-transformers==5.0.0
torch==2.8.0
numpy==2.3.0

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.20